tests/test_cpu.py
import unittest
from models.cpu import CPU
from models.memory import Memory
from models.alu import ALU
# Import other necessary components like Instruction if needed

class TestCPU(unittest.TestCase):

    def setUp(self):
        # Setup runs before each test method
        self.cpu = CPU()
        self.memory = Memory(1024)  # Assuming a memory size of 1024 for the tests
        self.alu = ALU()
        # Initialize CPU with ALU and memory, if the architecture of your CPU class requires it

    def test_cpu_initial_state(self):
        # Test the initial state of the CPU
        self.assertEqual(self.cpu.pc, 0, "Program counter should start at 0")
        self.assertFalse(self.cpu.halted, "CPU should not be halted at start")
        # Add more assertions for the initial state as needed

    def test_cpu_execute_instruction(self):
        # Test CPU's ability to execute a specific instruction
        # You would need a way to simulate an instruction. For instance:
        instruction = 0x20080001  # This is a MIPS instruction for 'addi $t0, $t0, 1'
        self.cpu.execute(instruction)
        # Check the effect of the instruction, e.g., register values, program counter increment, etc.
        # self.assertEqual(self.cpu.registers[8], 1, "Register $t0 should contain value 1")
        # self.assertEqual(self.cpu.pc, 4, "Program counter should increment")

    def test_memory_read_write(self):
        # Test reading and writing to memory
        test_address = 10
        test_value = 99
        self.memory.write(test_address, test_value)
        read_value = self.memory.read(test_address)
        self.assertEqual(test_value, read_value, f"Memory at address {test_address} should contain value {test_value}")

    # Add more tests to cover all aspects of CPU functionality
    # including edge cases and possible error conditions

if __name__ == '__main__':
    unittest.main()
